Done some Dagger modules refactoring - creating new modules and transferring some methods between them
