+++ app/src/main/java/com/aconno/sensorics/dagger/action/ActionScope.kt
@@ -1,0 +1,7 @@
+package com.aconno.sensorics.dagger.action
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ActionScope
