+++ app/src/main/java/com/aconno/sensorics/dagger/actionoutcome/ActionOutcomeScope.kt
@@ -1,0 +1,7 @@
+package com.aconno.sensorics.dagger.actionoutcome
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ActionOutcomeScope
