+++ app/src/main/java/com/aconno/sensorics/dagger/application/ActivityBuilder.kt
@@ -3,8 +3,12 @@ package com.aconno.sensorics.dagger.application
 import com.aconno.sensorics.AlarmService
 import com.aconno.sensorics.BluetoothConnectService
 import com.aconno.sensorics.BluetoothScanningService
-import com.aconno.sensorics.dagger.action_details.ActionDetailsActivityScope
-import com.aconno.sensorics.dagger.action_details.ActionDetailsModule
+import com.aconno.sensorics.dagger.action.ActionModule
+import com.aconno.sensorics.dagger.action.ActionScope
+import com.aconno.sensorics.dagger.actiondetails.ActionDetailsActivityScope
+import com.aconno.sensorics.dagger.actiondetails.ActionDetailsModule
+import com.aconno.sensorics.dagger.actionoutcome.ActionOutcomeModule
+import com.aconno.sensorics.dagger.actionoutcome.ActionOutcomeScope
 import com.aconno.sensorics.dagger.actionlist.ActionListFragmentsModule
 import com.aconno.sensorics.dagger.actionlist.ActionListModule
 import com.aconno.sensorics.dagger.actionlist.ActionListScope
@@ -14,6 +18,8 @@ import com.aconno.sensorics.dagger.bluetoothscanning.BluetoothScanningServiceMod
 import com.aconno.sensorics.dagger.bluetoothscanning.BluetoothScanningServiceScope
 import com.aconno.sensorics.dagger.configure.ConfigurationModule
 import com.aconno.sensorics.dagger.configure.ConfigurationScope
+import com.aconno.sensorics.dagger.device.DeviceModule
+import com.aconno.sensorics.dagger.device.DeviceScope
 import com.aconno.sensorics.dagger.deviceselect.DeviceSelectionFragmentsModule
 import com.aconno.sensorics.dagger.dfu.DfuActivityScope
 import com.aconno.sensorics.dagger.gcloudpublisher.GoogleCloudPublisherModule
@@ -23,9 +29,15 @@ import com.aconno.sensorics.dagger.mainactivity.MainActivityModule
 import com.aconno.sensorics.dagger.mainactivity.MainActivityScope
 import com.aconno.sensorics.dagger.mqttpublisher.MqttPublisherModule
 import com.aconno.sensorics.dagger.mqttpublisher.MqttPublisherScope
-import com.aconno.sensorics.dagger.publish.PublishListFragmentsModule
-import com.aconno.sensorics.dagger.publish.PublishListModule
-import com.aconno.sensorics.dagger.publish.PublishListScope
+import com.aconno.sensorics.dagger.publish.PublishModule
+import com.aconno.sensorics.dagger.publish.PublishScope
+import com.aconno.sensorics.dagger.publisher.PublisherModule
+import com.aconno.sensorics.dagger.publisher.PublisherScope
+import com.aconno.sensorics.dagger.publishlist.PublishListFragmentsModule
+import com.aconno.sensorics.dagger.publishlist.PublishListModule
+import com.aconno.sensorics.dagger.publishlist.PublishListScope
+import com.aconno.sensorics.dagger.readings.SensorReadingsModule
+import com.aconno.sensorics.dagger.readings.SensorReadingsScope
 import com.aconno.sensorics.dagger.restpublisher.RESTPublisherModule
 import com.aconno.sensorics.dagger.restpublisher.RESTPublisherScope
 import com.aconno.sensorics.dagger.splash.SplashActivityModule
@@ -47,13 +59,21 @@ import dagger.android.ContributesAndroidInjector
 abstract class ActivityBuilder {
 
     @MainActivityScope
+    @DeviceScope
+    @SensorReadingsScope
+    @ActionScope
     @ContributesAndroidInjector(
-        modules = [MainActivityModule::class, MainActivityFragmentsModule::class]
+        modules = [MainActivityModule::class, MainActivityFragmentsModule::class,
+        DeviceModule::class, SensorReadingsModule::class, ActionModule::class]
     )
     abstract fun bindMainActivity(): MainActivity
 
     @BluetoothScanningServiceScope
-    @ContributesAndroidInjector(modules = [BluetoothScanningServiceModule::class])
+    @SensorReadingsScope
+    @ActionOutcomeScope
+    @PublishScope
+    @ContributesAndroidInjector(modules = [BluetoothScanningServiceModule::class,
+        SensorReadingsModule::class,ActionOutcomeModule::class, PublishModule::class])
     abstract fun bindBluetoothScanningService(): BluetoothScanningService
 
     @ContributesAndroidInjector
@@ -66,30 +86,43 @@ abstract class ActivityBuilder {
     abstract fun bindPublishListActivity(): PublishListActivity
 
     @ActionListScope
+    @ActionScope
+    @DeviceScope
     @ContributesAndroidInjector(
-        modules = [ActionListModule::class, ActionListFragmentsModule::class]
+        modules = [ActionListModule::class, ActionListFragmentsModule::class, DeviceModule::class,
+            ActionModule::class]
     )
     abstract fun bindActionListActivity(): ActionListActivity
 
     @ActionDetailsActivityScope
-    @ContributesAndroidInjector(modules = [ActionDetailsModule::class])
+    @ActionScope
+    @ContributesAndroidInjector(modules = [ActionDetailsModule::class, ActionModule::class])
     abstract fun bindActionDetailsActivity(): ActionDetailsActivity
 
     @GoogleCloudPublisherScope
+    @PublishScope
+    @PublisherScope
     @ContributesAndroidInjector(
-        modules = [GoogleCloudPublisherModule::class, DeviceSelectionFragmentsModule::class]
+        modules = [GoogleCloudPublisherModule::class, DeviceSelectionFragmentsModule::class,
+        PublishModule::class,PublisherModule::class]
     )
     abstract fun bindGoogleCloudPublisherActivity(): GoogleCloudPublisherActivity
 
     @MqttPublisherScope
+    @PublishScope
+    @PublisherScope
     @ContributesAndroidInjector(
-        modules = [MqttPublisherModule::class, DeviceSelectionFragmentsModule::class]
+        modules = [MqttPublisherModule::class, DeviceSelectionFragmentsModule::class,
+            PublishModule::class,PublisherModule::class]
     )
     abstract fun bindMqttPublisherActivity(): MqttPublisherActivity
 
     @RESTPublisherScope
+    @PublishScope
+    @PublisherScope
     @ContributesAndroidInjector(
-        modules = [RESTPublisherModule::class, DeviceSelectionFragmentsModule::class]
+        modules = [RESTPublisherModule::class, DeviceSelectionFragmentsModule::class,
+            PublishModule::class,PublisherModule::class]
     )
     abstract fun bindRestPublisherActivity(): RestPublisherActivity
