+++ app/src/main/java/com/aconno/sensorics/dagger/bluetoothscanning/BluetoothScanningServiceModule.kt
@@ -59,104 +59,6 @@ class BluetoothScanningServiceModule {
     @BluetoothScanningServiceScope
     fun provideBluetoothScanningReceiverIntentFilter() = IntentFilter("com.aconno.sensorics.STOP")
 
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideRecordSensorValuesUseCase(
-        inMemoryRepository: InMemoryRepository
-    ): SaveSensorReadingsUseCase {
-        return SaveSensorReadingsUseCase(inMemoryRepository)
-    }
 
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideLogReadingsUseCase(
-        bluetoothScanningService: BluetoothScanningService
-    ): LogReadingUseCase {
-        return LogReadingUseCase(FileStorageImpl(bluetoothScanningService))
-    }
 
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideHandleInputUseCase(
-        actionsRepository: ActionsRepository,
-        getLocalTimeOfDayInSecondsUseCase: GetLocalTimeOfDayInSecondsUseCase
-    ): InputToOutcomesUseCase {
-        return InputToOutcomesUseCase(actionsRepository, getLocalTimeOfDayInSecondsUseCase)
-    }
-
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideGetLocalTimeOfDayInSecondsUseCase(
-        timeProvider: TimeProvider
-    ): GetLocalTimeOfDayInSecondsUseCase {
-        return GetLocalTimeOfDayInSecondsUseCase(timeProvider)
-    }
-
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideRunOutcomeUseCase(
-        notificationDisplay: NotificationDisplay,
-        textToSpeechPlayer: TextToSpeechPlayer,
-        vibrator: Vibrator,
-        alarmServiceController: AlarmServiceController
-    ): RunOutcomeUseCase {
-        val notificationOutcomeExecutor = NotificationOutcomeExecutor(notificationDisplay)
-        val textToSpeechOutcomeExecutor = TextToSpeechOutcomeExecutor(textToSpeechPlayer)
-        val vibrationOutcomeExecutor = VibrationOutcomeExecutor(vibrator)
-        val alarmOutcomeExecutor = AlarmOutcomeExecutor(alarmServiceController)
-        val outcomeExecutorSelector = OutcomeExecutorSelector(
-            notificationOutcomeExecutor,
-            textToSpeechOutcomeExecutor,
-            vibrationOutcomeExecutor,
-            alarmOutcomeExecutor
-        )
-
-        return RunOutcomeUseCase(outcomeExecutorSelector)
-    }
-
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideGetAllEnabledGooglePublishUseCase(
-        googlePublishRepository: GooglePublishRepository
-    ): GetAllEnabledGooglePublishUseCase {
-        return GetAllEnabledGooglePublishUseCase(
-            googlePublishRepository
-        )
-    }
-
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideGetAllEnabledRESTPublishUseCase(
-        restPublishRepository: RestPublishRepository
-    ): GetAllEnabledRestPublishUseCase {
-        return GetAllEnabledRestPublishUseCase(
-            restPublishRepository
-        )
-    }
-
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideGetAllEnabledMqttPublishUseCase(
-        mqttPublishRepository: MqttPublishRepository
-    ): GetAllEnabledMqttPublishUseCase {
-        return GetAllEnabledMqttPublishUseCase(
-            mqttPublishRepository
-        )
-    }
-
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideGetRESTHeadersByIdUseCase(
-        restPublishRepository: RestPublishRepository
-    ): GetRestHeadersByIdUseCase {
-        return GetRestHeadersByIdUseCase(restPublishRepository)
-    }
-
-    @Provides
-    @BluetoothScanningServiceScope
-    fun provideGetRESTHttpGetParamsByIdUseCase(
-        restPublishRepository: RestPublishRepository
-    ): GetRestHttpGetParamsByIdUseCase {
-        return GetRestHttpGetParamsByIdUseCase(restPublishRepository)
-    }
 }
