+++ app/src/main/java/com/aconno/sensorics/dagger/device/DeviceScope.kt
@@ -1,0 +1,7 @@
+package com.aconno.sensorics.dagger.device
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class DeviceScope
