+++ app/src/main/java/com/aconno/sensorics/dagger/mainactivity/MainActivityModule.kt
@@ -3,7 +3,6 @@ package com.aconno.sensorics.dagger.mainactivity
 import androidx.lifecycle.ViewModelProviders
 import com.aconno.sensorics.BluetoothStateReceiver
 import com.aconno.sensorics.SensoricsApplication
-import com.aconno.sensorics.dagger.action_details.ActionDetailsActivityScope
 import com.aconno.sensorics.device.permissons.PermissionActionFactory
 import com.aconno.sensorics.domain.actions.ActionsRepository
 import com.aconno.sensorics.domain.interactor.filter.FilterByMacUseCase
@@ -115,29 +114,6 @@ class MainActivityModule {
         bluetoothViewModelFactory
     ).get(BluetoothViewModel::class.java)
 
-    @Provides
-    @MainActivityScope
-    fun provideGetAllDevicesUseCase(
-        deviceRepository: DeviceRepository
-    ): GetSavedDevicesUseCase {
-        return GetSavedDevicesUseCase(deviceRepository)
-    }
-
-    @Provides
-    @MainActivityScope
-    fun provideSaveDeviceUseCase(
-        deviceRepository: DeviceRepository
-    ): SaveDeviceUseCase {
-        return SaveDeviceUseCase(deviceRepository)
-    }
-
-    @Provides
-    @MainActivityScope
-    fun provideDeleteDeviceUseCase(
-        deviceRepository: DeviceRepository
-    ): DeleteDeviceUseCase {
-        return DeleteDeviceUseCase(deviceRepository)
-    }
 
     @Provides
     @MainActivityScope
@@ -226,12 +202,6 @@ class MainActivityModule {
 
     @Provides
     @MainActivityScope
-    fun provideGetSensorReadingsUseCase(
-        inMemoryRepository: InMemoryRepository
-    ) = GetReadingsUseCase(inMemoryRepository)
-
-    @Provides
-    @MainActivityScope
     fun provideMainResourceViewModelFactory(
         getMainResourceUseCase: GetMainResourceUseCase
     ) = MainResourceViewModelFactory(getMainResourceUseCase)
@@ -244,24 +214,5 @@ class MainActivityModule {
     ) = ViewModelProviders.of(mainActivity, mainResourceViewModelFactory)
         .get(MainResourceViewModel::class.java)
 
-    @Provides
-    @MainActivityScope
-    fun provideAddActionUseCase(
-        actionsRepository: ActionsRepository
-    ) = AddActionUseCase(actionsRepository)
 
-    @Provides
-    @MainActivityScope
-    fun provideGetActionsByDeviceMacAddressUseCase(
-        actionsRepository: ActionsRepository
-    ) = GetActionsByDeviceMacAddressUseCase(actionsRepository)
-
-    @Provides
-    @MainActivityScope
-    fun provideSetActionActiveByDeviceMacAddressUseCase(
-        addActionUseCase: AddActionUseCase,
-        getActionsByDeviceMacAddressUseCase: GetActionsByDeviceMacAddressUseCase
-    ) = SetActionActiveByDeviceMacAddressUseCase(
-        addActionUseCase, getActionsByDeviceMacAddressUseCase
-    )
 }
