+++ app/src/main/java/com/aconno/sensorics/dagger/publisher/PublisherScope.kt
@@ -1,0 +1,7 @@
+package com.aconno.sensorics.dagger.publisher
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class PublisherScope
