+++ app/src/main/java/com/aconno/sensorics/dagger/readings/SensorReadingsScope.kt
@@ -1,0 +1,7 @@
+package com.aconno.sensorics.dagger.readings
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class SensorReadingsScope
